Add GetARoundMode and SetRoundMode to Backwards. Changed the conversion to double to hopefully be faster without compromising correctness. Fixed a bug in conversion from double that was causing an unnecessary rounding (that still seems to be occurring for some reason).
First note: I thought about implementing the rounding functions natively, but I can't think of a case where they would be incorrect, so long as the SlowFloat is correctly converted to double.
Second note: Need to figure out why rounding is still broken for rounding to positive infinity on Windows.